requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.3.0=mkl
abseil-cpp=20211102.0=hd77b12b_0
absl-py=1.4.0=py310haa95532_0
aiohttp=3.8.5=py310h2bbff1b_0
aiosignal=1.2.0=pyhd3eb1b0_0
astunparse=1.6.3=py_0
async-timeout=4.0.2=py310haa95532_0
attrs=22.1.0=py310haa95532_0
blas=1.0=mkl
blinker=1.4=py310haa95532_0
brotlipy=0.7.0=py310h2bbff1b_1002
bzip2=1.0.8=he774522_0
c-ares=1.19.1=h2bbff1b_0
ca-certificates=2023.08.22=haa95532_0
cachecontrol=0.13.1=pypi_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2023.7.22=py310haa95532_0
cffi=1.15.1=py310h2bbff1b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.3=pyhd3eb1b0_0
cmake=3.27.6=pypi_0
colorama=0.4.6=pypi_0
cryptography=41.0.3=py310h3438e0d_0
deprecated=1.2.14=pypi_0
distlib=0.3.7=pypi_0
filelock=3.13.1=pypi_0
firebase-admin=6.2.0=pypi_0
flask=2.2.2=py310haa95532_0
flatbuffers=2.0.0=h6c2663c_0
frozenlist=1.3.3=py310h2bbff1b_0
gast=0.4.0=pyhd3eb1b0_0
gcloud=0.18.3=pypi_0
giflib=5.2.1=h8cc25b3_3
google-api-core=2.12.0=pypi_0
google-api-python-client=2.102.0=pypi_0
google-auth=2.22.0=py310haa95532_0
google-auth-httplib2=0.1.1=pypi_0
google-auth-oauthlib=0.4.4=pyhd3eb1b0_0
google-cloud-core=2.3.3=pypi_0
google-cloud-firestore=2.12.0=pypi_0
google-cloud-storage=2.11.0=pypi_0
google-crc32c=1.5.0=pypi_0
google-pasta=0.2.0=pyhd3eb1b0_0
google-resumable-media=2.6.0=pypi_0
googleapis-common-protos=1.60.0=pypi_0
grpc-cpp=1.48.2=hf108199_0
grpcio=1.59.0=pypi_0
grpcio-status=1.59.0=pypi_0
h5py=3.9.0=py310hfc34f40_0
hdf5=1.12.1=h51c971a_3
httplib2=0.22.0=pypi_0
icc_rt=2022.1.0=h6049295_2
icu=58.2=ha925a31_3
idna=3.4=py310haa95532_0
intel-openmp=2023.1.0=h59b6b97_46319
itsdangerous=2.0.1=pyhd3eb1b0_0
jinja2=3.1.2=py310haa95532_0
joblib=1.2.0=py310haa95532_0
jpeg=9e=h2bbff1b_1
jsonify=0.5=pypi_0
jwcrypto=1.5.0=pypi_0
keras=2.10.0=py310haa95532_0
keras-preprocessing=1.1.2=pyhd3eb1b0_0
libcurl=8.2.1=h86230a5_0
libffi=3.4.4=hd77b12b_0
libpng=1.6.39=h8cc25b3_0
libprotobuf=3.20.3=h23ce68f_0
libssh2=1.10.0=hcd4344a_2
markdown=3.4.1=py310haa95532_0
markupsafe=2.1.1=py310h2bbff1b_0
mkl=2023.1.0=h6b88ed4_46357
mkl-service=2.4.0=py310h2bbff1b_1
mkl_fft=1.3.8=py310h2bbff1b_0
mkl_random=1.2.4=py310h59b6b97_0
msgpack=1.0.7=pypi_0
multidict=6.0.2=py310h2bbff1b_0
nltk=3.8.1=py310haa95532_0
numpy=1.25.2=py310h055cbcc_0
numpy-base=1.25.2=py310h65a83cf_0
oauth2client=4.1.3=pypi_0
oauthlib=3.2.2=py310haa95532_0
openssl=1.1.1w=h2bbff1b_0
opt_einsum=3.3.0=pyhd3eb1b0_1
packaging=23.1=py310haa95532_0
pip=23.2.1=py310haa95532_0
platformdirs=3.11.0=pypi_0
proto-plus=1.22.3=pypi_0
protobuf=3.20.3=pypi_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd3eb1b0_0
pycryptodome=3.19.0=pypi_0
pyjwt=2.8.0=pypi_0
pyopenssl=23.2.0=py310haa95532_0
pyparsing=3.1.1=pypi_0
pyrebase4=4.7.1=pypi_0
pysocks=1.7.1=py310haa95532_0
python=3.10.0=h96c0403_3
python-flatbuffers=2.0=pyhd3eb1b0_0
python-jwt=4.0.0=pypi_0
re2=2022.04.01=hd77b12b_0
regex=2022.7.9=py310h2bbff1b_0
requests=2.29.0=pypi_0
requests-oauthlib=1.3.0=py_0
requests-toolbelt=0.10.1=pypi_0
rsa=4.7.2=pyhd3eb1b0_1
sastrawi=1.0.1=pypi_0
scipy=1.11.1=py310h309d312_0
setuptools=68.0.0=py310haa95532_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.9=h6c2663c_0
sqlite=3.41.2=h2bbff1b_0
tbb=2021.8.0=h59b6b97_0
tensorboard=2.10.0=py310haa95532_0
tensorboard-data-server=0.6.1=py310haa95532_0
tensorboard-plugin-wit=1.8.1=py310haa95532_0
tensorflow=2.10.0=mkl_py310hd99672f_0
tensorflow-base=2.10.0=mkl_py310h6a7f48e_0
tensorflow-estimator=2.10.0=py310haa95532_0
termcolor=2.1.0=py310haa95532_0
tk=8.6.12=h2bbff1b_0
tqdm=4.62.3=pyhd3eb1b0_1
typing_extensions=4.7.1=py310haa95532_0
tzdata=2023c=h04d1e81_0
uritemplate=4.1.1=pypi_0
urllib3=1.26.16=py310haa95532_0
vc=14.2=h21ff451_1
virtualenv=20.24.6=pypi_0
vs2015_runtime=14.27.29016=h5e58377_2
werkzeug=2.2.3=py310haa95532_0
wheel=0.38.4=py310haa95532_0
win_inet_pton=1.1.0=py310haa95532_0
wrapt=1.14.1=py310h2bbff1b_0
xz=5.4.2=h8cc25b3_0
yarl=1.8.1=py310h2bbff1b_0
zlib=1.2.13=h8cc25b3_0
